Add Cypress component testing for lit
