+++ cypress/fixtures/profile.json
@@ -1,0 +1,5 @@
+{
+  "id": 8739,
+  "name": "Jane",
+  "email": "jane@example.com"
+}
