+++ package.json
@@ -61,6 +61,7 @@
     "concurrently": "9.2.1",
     "cssnano": "7.1.2",
     "cypress": "15.10.0",
+    "cypress-ct-lit": "1.0.0",
     "esbuild": "0.27.3",
     "esbuild-plugin-liquid": "0.2.0",
     "esbuild-plugin-minify-html-literals": "3.0.0",
